projects/37_langchain_rag/pyproject.toml
[tool.poetry]
name = "langchain-rag"
version = "0.1.0"
description = ""
authors = ["Viktor Szabo <szabo_viktor@lilly.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
gradio = ">=5.38.0,<6.0.0"
langchain = ">=0.3.26,<0.4.0"
langchain-ibm = ">=0.3.15,<0.4.0"
langchain-community = ">=0.3.27,<0.4.0"
langchain-text-splitters = ">=0.3.8,<0.4.0"
python-dotenv = ">=1.1.1,<2.0.0"
ibm-watsonx-ai = ">=1.3.31,<2.0.0"
chromadb = ">=1.0.15,<2.0.0"
pypdf = ">=5.8.0,<6.0.0"
pymupdf = ">=1.23.0,<2.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
